requirements.txt
beautifulsoup4==4.12.2
html2text==2020.1.16
itchat_uos==1.5.0.dev0
python-dotenv==1.0.0
Requests==2.31.0
selenium==4.15.2
xmltodict==0.13.0
zhipuai==1.0.7
